ref: allow MissingFeatureError from createJWT
