+++ src/app/modules/spcp/spcp.types.ts
@@ -25,3 +25,8 @@ export interface CorppassAttributes {
     CPUID?: string
   }
 }
+
+export interface ExtractedAttributes {
+  userName: string
+  userInfo?: string
+}
